Updated with recent changes including Dropdown support, better support for Checkboxes, future support for fancy text, better input handling;

+++ Development/Source/Future.cpp
@@ -18,7 +18,7 @@ void Future::Update()
 	{
 		Found = false;
 
-		for(size_t i = Futures.size() - 1; i >= 0; i--)
+		for(long i = Futures.size() - 1; i >= 0; i--)
 		{
 			if(time(NULL) - Futures[i].StartTime >= Futures[i].Length)
 			{
